requirements.txt
# Core MCP framework
fastmcp>=0.1.0

# Email processing
aiofiles>=0.8.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.991